Minor code tweaks/tidy and including pull request #149

+++ app/src/main/java/com/garethevans/church/opensongtablet/filemanagement/StorageAccess.java
@@ -572,7 +572,7 @@ public class StorageAccess {
         // Get rid of all the uri info up to the end of /OpenSong/Songs
         // Also adds mainfoldername if the song isn't in a subfolder
         if (uriString.contains("OpenSong/Songs/")) {
-            uriString = uriString.substring(uriString.indexOf("OpenSong/Songs/") + 15);
+            uriString = uriString.substring(uriString.lastIndexOf("OpenSong/Songs/") + 15);
         }
         if (!uriString.contains("/")) {
             uriString = mainfolder + "/" + uriString;
@@ -1497,8 +1497,8 @@ public class StorageAccess {
             f = new File(f, filename);
             Uri uri = Uri.fromFile(f);
             try {
-                if (uriExists(uri)) {
-                    // IV - Delete any old file
+                if (f.exists() && uriExists(uri)) {
+                    // IV - Delete any existing file (does not touch folder)
                     Log.d(TAG,"Removing preexisting file: filename - "+ f.delete());
                 }
                 stuffCreated = f.createNewFile();
